WIP: refactor field tests to support rings too
