+++ icicle/cmake/target_editor.cmake
@@ -18,10 +18,11 @@ function(handle_curve TARGET)
 endfunction()
 
 function(handle_ring TARGET)
-  # TODO: add ring specific sources
-  # target_sources(${TARGET} PRIVATE
-  #     src/vec_ops.cpp  
-  # )
+  target_sources(${TARGET} PRIVATE
+    src/fields/ffi_extern.cpp
+    src/vec_ops.cpp
+    src/matrix_ops.cpp
+  )
 endfunction()
 
 function(handle_ntt TARGET FEATURE_LIST)
