+++ icicle/tests/CMakeLists.txt
@@ -76,7 +76,7 @@ endif()
 if (RING)
     add_executable(test_ring_api test_ring_api.cpp)
     target_include_directories(test_ring_api PRIVATE ${CMAKE_SOURCE_DIR}/include/)
-    target_link_libraries(test_ring_api PRIVATE GTest::gtest_main icicle_device)
+    target_link_libraries(test_ring_api PRIVATE GTest::gtest_main icicle_device icicle_ring)
     # TODO Yuval: target_link_libraries(test_ring_api PRIVATE icicle_ring)
     gtest_discover_tests(test_ring_api)
 endif()
